Extract matrix relatec code to own class
