+++ src/request_collector/server.py
@@ -1,34 +1,14 @@
-import json
 import sys
 import traceback
-from typing import Iterable, List, Tuple
 
 import gevent
 import structlog
-from eth_utils import decode_hex, encode_hex, to_checksum_address
-from matrix_client.errors import MatrixRequestError
+from eth_utils import encode_hex, to_checksum_address
+from request_collector.matrix import MatrixListener
 
 from monitoring_service.database import SharedDatabase
 from monitoring_service.states import MonitorRequest
-from raiden.constants import MONITORING_BROADCASTING_ROOM, Environment
-from raiden.exceptions import InvalidProtocolMessage, TransportError
-from raiden.messages import RequestMonitoring, SignedMessage, from_dict as message_from_dict
-from raiden.network.transport.matrix.client import GMatrixClient, Room
-from raiden.network.transport.matrix.utils import (
-    join_global_room,
-    login_or_register,
-    make_client,
-    make_room_alias,
-    validate_userid_signature,
-)
-from raiden.network.transport.udp import udp_utils
-from raiden.settings import (
-    DEFAULT_MATRIX_KNOWN_SERVERS,
-    DEFAULT_TRANSPORT_MATRIX_RETRY_INTERVAL,
-    DEFAULT_TRANSPORT_RETRIES_BEFORE_BACKOFF,
-)
-from raiden.utils.cli import get_matrix_servers
-from raiden.utils.signer import LocalSigner
+from raiden.messages import RequestMonitoring, SignedMessage
 from raiden_libs.exceptions import InvalidSignature
 from raiden_libs.gevent_error_handler import register_error_handler
 
@@ -45,47 +25,6 @@ def error_handler(_context, exc_info):
     sys.exit()
 
 
-def setup_matrix(private_key: str, chain_id: int) -> Tuple[GMatrixClient, Room]:
-    available_servers_url = DEFAULT_MATRIX_KNOWN_SERVERS[Environment.DEVELOPMENT]
-    available_servers = get_matrix_servers(available_servers_url)
-
-    def _http_retry_delay() -> Iterable[float]:
-        # below constants are defined in raiden.app.App.DEFAULT_CONFIG
-        return udp_utils.timeout_exponential_backoff(
-            DEFAULT_TRANSPORT_RETRIES_BEFORE_BACKOFF,
-            DEFAULT_TRANSPORT_MATRIX_RETRY_INTERVAL / 5,
-            DEFAULT_TRANSPORT_MATRIX_RETRY_INTERVAL,
-        )
-
-    client = make_client(
-        servers=available_servers,
-        http_pool_maxsize=4,
-        http_retry_timeout=40,
-        http_retry_delay=_http_retry_delay,
-    )
-
-    try:
-        login_or_register(client, signer=LocalSigner(private_key=decode_hex(private_key)))
-    except MatrixRequestError:
-        raise ConnectionError('Could not login/register to matrix.')
-    except ValueError:
-        raise ConnectionError('Could not login/register to matrix.')
-
-    try:
-        room_name = make_room_alias(chain_id, MONITORING_BROADCASTING_ROOM)
-        monitoring_room = join_global_room(
-            client=client,
-            name=room_name,
-            servers=available_servers,
-        )
-    except MatrixRequestError:
-        raise ConnectionError('Could not join monitoring broadcasting room.')
-    except TransportError:
-        raise ConnectionError('Could not join monitoring broadcasting room.')
-
-    return client, monitoring_room
-
-
 class RequestCollector(gevent.Greenlet):
     def __init__(
         self,
@@ -99,11 +38,11 @@ class RequestCollector(gevent.Greenlet):
 
         state = self.state_db.load_state(0)
         try:
-            self.client, self.monitoring_room = setup_matrix(
-                private_key=self.private_key,
+            self.matrix_listener = MatrixListener(
+                private_key=private_key,
                 chain_id=state.blockchain_state.chain_id,
+                callback=self.handle_message,
             )
-            self.monitoring_room.add_listener(self._handle_message, 'm.room.message')
         except ConnectionError as e:
             log.critical(
                 'Could not connect to broadcasting system.',
@@ -114,99 +53,13 @@ class RequestCollector(gevent.Greenlet):
     def _run(self):
         register_error_handler(error_handler)
 
-        self.client.start_listener_thread()
-        self.client.sync_thread.get().wait()
+        self.matrix_listener.start()
 
     def stop(self):
-        self.client.stop_listener_thread()
+        self.matrix_listener.stop()
+        self.matrix_listener.join()
 
-    def _handle_message(self, room, event) -> bool:
-        """ Handle text messages sent to listening rooms """
-        if (
-                event['type'] != 'm.room.message' or
-                event['content']['msgtype'] != 'm.text'
-        ):
-            # Ignore non-messages and non-text messages
-            return False
-
-        sender_id = event['sender']
-
-        user = self.client.get_user(sender_id)
-        peer_address = validate_userid_signature(user)
-        if not peer_address:
-            log.debug(
-                'Message from invalid user displayName signature',
-                peer_user=user.user_id,
-                room=room,
-            )
-            return False
-
-        data = event['content']['body']
-        if not isinstance(data, str):
-            log.warning(
-                'Received message body not a string',
-                peer_user=user.user_id,
-                peer_address=to_checksum_address(peer_address),
-                room=room,
-            )
-            return False
-
-        messages: List[SignedMessage] = list()
-
-        for line in data.splitlines():
-            line = line.strip()
-            if not line:
-                continue
-            try:
-                message_dict = json.loads(line)
-                message = message_from_dict(message_dict)
-            except (UnicodeDecodeError, json.JSONDecodeError) as ex:
-                log.warning(
-                    "Can't parse message data JSON",
-                    message_data=line,
-                    peer_address=to_checksum_address(peer_address),
-                    _exc=ex,
-                )
-                continue
-            except InvalidProtocolMessage as ex:
-                log.warning(
-                    "Message data JSON are not a valid message",
-                    message_data=line,
-                    peer_address=to_checksum_address(peer_address),
-                    _exc=ex,
-                )
-                continue
-            if not isinstance(message, SignedMessage):
-                log.warning(
-                    'Received invalid message',
-                    message=message,
-                )
-                continue
-            elif message.sender != peer_address:
-                log.warning(
-                    'Message not signed by sender!',
-                    message=message,
-                    signer=message.sender,
-                    peer_address=peer_address,
-                )
-                continue
-            messages.append(message)
-
-        if not messages:
-            return False
-
-        for message in messages:
-            self._receive_message(message)
-
-        return True
-
-    def _receive_message(self, message: SignedMessage):
-        log.debug(
-            'Message received',
-            message=message,
-            sender=to_checksum_address(message.sender),
-        )
-
+    def handle_message(self, message: SignedMessage):
         if isinstance(message, RequestMonitoring):
             self.on_monitor_request(message)
         else:
@@ -233,13 +86,6 @@ class RequestCollector(gevent.Greenlet):
                 non_closing_signature=encode_hex(request_monitoring.non_closing_signature),
                 reward_amount=request_monitoring.reward_amount,
                 reward_proof_signature=encode_hex(request_monitoring.signature),
-            )
-            with self.state_db.conn:
-                self.state_db.upsert_monitor_request(monitor_request)
-
-            log.debug(
-                'Stored new monitor request in database',
-                request=monitor_request,
             )
         except InvalidSignature:
             log.info('Ignore MR with invalid signature {}'.format(request_monitoring))
